@@ -1,39 +1,44 @@
-import React, { Component } from 'react';
+import React, { useState } from 'react';
 import './App.css';
 import Person from './Person/Person';
 
-class App extends Component {
-  state = {
-    person:[
+const app = props => {
+  const [personsState, setPersonsState] = useState({
+    persons:[
       {name:'Max', age:30},
       {name:'Manu', age:33},
       {name:'Rex', age:32}
-    ]
-  }
-  switchNameHandler = () =>{
+    ],
+    otherProperty: 'This is another property'
+  });
+
+  const [otherState, setOtherState] = useState('some other value');
+  console.log(personsState, otherState);
+  const switchNameHandler = () => {
    //console.log('clicked!!');
-   //this.person[0].name = 'Maximilian'; DO NOT DO This
-   this.setState({
-     person:[
+   //this.persons[0].name = 'Maximilian'; DO NOT DO This
+   setPersonsState({
+     persons:[
        {name:'Maximilian', age:30},
        {name:'Manu', age:33},
        {name:'Rex', age:28}
-     ]
+     ],
+     otherProperty: 'This is another property'
    });
-  }
-  render() {
-    return (
-      <div className="App">
-        <h1>Hi, I'm a React app</h1>
-        <p>This is really working!</p>
-        <button onClick={this.switchNameHandler}>Click on me!</button>
-        <Person name={this.state.person[0].name} age={this.state.person[0].age}></Person>
-        <Person name={this.state.person[1].name} age={this.state.person[1].age}></Person>
-        <Person name={this.state.person[2].name} age={this.state.person[2].age}>My Hobbie is: soccer</Person>
-      </div>
-    );
-    //return React.createElement('div', {className: 'App'}, React.createElement('h1', null, 'Does it work now!'));
-  }
-}
 
-export default App;
+ };
+  return (
+    <div className="App">
+      <h1>Hi, I'm a React app</h1>
+      <p>This is really working!</p>
+      <button onClick={switchNameHandler}>Click on me!</button>
+      <Person name={personsState.persons[0].name} age={personsState.persons[0].age}></Person>
+      <Person name={personsState.persons[1].name} age={personsState.persons[1].age}></Person>
+      <Person name={personsState.persons[2].name} age={personsState.persons[2].age}>My Hobbie is: soccer</Person>
+    </div>
+  );
+    //return React.createElement('div', {className: 'App'}, React.createElement('h1', null, 'Does it work now!'));
+
+};
+
+export default app;
