@@ -1,12 +1,12 @@
 import React from 'react';
-const person = (props) => {
-  return (
-    <div>
-      <p>I'm {props.name} and I have {props.age} years old </p>
-      <p>{props.children}</p>
-    </div>
 
-  )
+const person = ( props ) => {
+    return (
+        <div>
+            <p>I'm {props.name} and I am {props.age} years old!</p>
+            <p>{props.children}</p>
+        </div>
+    )
 };
 
 export default person;
